feat(types): update responses for new api
