+++ src/lib/types/common.ts
@@ -1,0 +1,11 @@
+/**
+ *  Options for the Imperial class
+ */
+export interface ImperialOptions {
+	requestTimeout?: number;
+}
+
+/**
+ *  Id or url of the document
+ */
+export type IdResolvable = string | any;
